requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
motor[srv]
